NSLookup/requirements.txt
cortexutils
requests
dnspython